Refactor file explorer
