+++ frontend/src/routes/_oh.app._index/constants.ts
@@ -1,0 +1,11 @@
+export const ASSET_FILE_TYPES = [
+  ".png",
+  ".jpg",
+  ".jpeg",
+  ".bmp",
+  ".gif",
+  ".pdf",
+  ".mp4",
+  ".webm",
+  ".ogg",
+];
